+++ Views/Recepsion/Create.cshtml
@@ -14,7 +14,7 @@
 @model Usuario
 <div class="container ">
     <div class="row d-flex justify-content-center align-items-center">
-        <form asp-controller="Usuarios" asp-action="Add" method="post" class="card col-8 mt-3">
+        <form asp-controller="Recepsion" asp-action="Add" method="post" class="card col-8 mt-3">
             <div class="card-body">
                 <div class="form-grup mb-4 row">
                     <div class="col-6">
